+++ packages/frameworks/react/CHANGELOG.md
@@ -6,6 +6,14 @@ description: All notable changes to this project will be documented in this file
 
 ## [Unreleased]
 
+### Added
+
+- Added `ValueText` to the `ColorPicker` component
+
+### Fixed
+
+- Added missing data attributes to `DatePickerView` component
+
 ## [1.0.0-beta.1] - 2023-10-20
 
 ### Added