+++ packages/frameworks/vue/CHANGELOG.md
@@ -8,6 +8,7 @@ description: All notable changes to this project will be documented in this file
 
 ### Added
 
+- Added `ValueText` to the `ColorPicker` component
 - Added `ToggleGroup` component
 
 ### Changed